\docType{package}
\name{Rvisdiff}
\alias{Rvisdiff}
\title{Rvisdiff: Interactive Graphs for Differential Expression}
\description{
Rvisdiff creates a muti-graph web page which allows the interactive exploration
of differential expression results. The graphical web interface presents results
as a table which is integrated with five interactive graphs: MA-plot, volcano
plot, box plot, lines plot and cluster heatmap. Graphical aspect and information
represented in the graphs can be customized by means of user controls. Final
graphics can be exported as PNG format.
}
\details{
The main function is:
\itemize{
\item \code{\link{DEreport}} - produce interactive graphs of the given
differential expression results
}

For detailed information on usage, see the package vignette, by typing
\code{vignette("Rvisdiff")}.

The GitHub repository is \url{https://github.com/BioinfoUSAL/Rvisdiff}.
This is the place to file an issue, report a bug, or provide a pull request.
}
\author{
David Barrios and Carlos Prieto
}
